Create CustomElasticSearchMapping trait
